requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiolimiter==1.2.1
aiortc==1.13.0
aiosignal==1.4.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.41.5
botocore==1.41.5
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.2
ctranslate2==4.6.0
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.2.2
ddgs==9.9.2
decorator==4.4.2
defusedxml==0.7.1
demucs==4.0.1
dingpy==1.0.6
distro==1.9.0
dnspython==2.8.0
docutils==0.15.2
dora_search==0.1.12
dtw-python==1.7.2
duckduckgo_search==8.1.1
einops==0.8.1
fake-useragent==2.2.0
Faker==37.8.0
fastapi==0.117.1
faster-whisper==1.2.0
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.0
frozendict==2.4.7
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-crc32c==1.7.1
google-genai==1.51.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.72.0
gradio==5.46.1
gradio_client==1.13.1
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
kiwisolver==1.4.9
lameenc==1.8.1
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.30
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.45.0
lxml==6.0.2
macholib==1.16.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multitasking==0.0.12
narwhals==2.5.0
networkx==3.5
noisereduce==3.0.3
numba==0.62.0
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.0
openai==2.8.0
openai-whisper==20250625
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openunmix==1.3.0
orjson==3.11.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
peewee==3.18.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
primp==0.15.0
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.24.1
pyee==13.0.0
pyfiglet==1.0.4
pygame==2.6.1
pygbag==0.9.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyInputPlus==0.2.12
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pylibsrtp==0.12.0
PyMatting==1.1.14
PyMsgBox==2.0.1
PyMuPDF==1.26.4
pyngrok==7.3.0
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySimpleValidate==0.2.12
PySocks==1.7.1
pytesseract==0.3.13
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
rembg==2.0.67
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
rubicon-objc==0.5.3
ruff==0.13.1
s3transfer==0.15.0
safehttpx==0.1.6
safetensors==0.6.2
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.36.0
semantic-version==2.10.0
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SpeechRecognition==3.14.3
speedtest-cli==2.1.3
starlette==0.48.0
stdiomask==0.0.6
streamlit==1.49.1
streamlit-cropper==0.3.1
streamlit-mic-recorder==0.0.4
streamlit-webrtc==0.63.11
submitit==1.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
treetable==0.2.6
trio==0.31.0
trio-websocket==0.12.2
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
voyageai==0.3.5
watchdog==6.0.0
wavio==0.0.9
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
yfinance==0.2.66
youtube-comment-downloader==0.1.78
youtube-transcript-api==1.2.3
yt-dlp==2025.9.26
zstandard==0.25.0
